requirements.txt
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
google-api-core==2.29.0
google-auth==2.48.0
google-cloud-bigquery==3.40.0
google-cloud-core==2.5.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
idna==3.11
numpy==2.4.2
packaging==26.0
pandas==2.2.3
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
beautifulsoup4==4.12.3
tenacity==9.1.2
db-dtypes==1.5.0